enhance: support a remote grapher server for deploys
Previously, it was assumed that the live server for Grapher was the same
server that the covid scripts were running on. In that scenario,
deploying just means writing to a file.

This change means that the live server can be on a different machine,
and it will ssh to that machine and set the deploy queue path.